requirements-dev.txt
-r requirements.txt
einops
graphviz
mypy
mako
oyaml
parameterized
pre-commit
pybind11
pybind11-stubgen
pytest
pytest-asyncio
pytest-cov
pytest-csv
pytest-env
pytest-forked
pytest-xdist
pytest-timeout
pytest-split
pytest-mock
pytest-threadleak
rouge_score
cloudpickle
typing-extensions==4.12.2
bandit==1.7.7
jsonlines==4.0.0
jieba==0.42.1
rouge==1.0.1
pytest-rerunfailures
ruff==0.9.4
lm_eval[api]==0.4.10
docstring_parser
genai-perf==0.0.13
opentelemetry-sdk>=1.26.0
opentelemetry-api>=1.26.0
opentelemetry-exporter-otlp>=1.26.0
opentelemetry-semantic-conventions-ai>=0.4.1
fuzzywuzzy==0.18.0
aiperf==0.3.0
nanobind>=2.9.0
nixl==0.8.0
hf-transfer==0.1.9
line_profiler
